+++ Source/TitleScrollView.swift
@@ -29,7 +29,12 @@ public protocol TitleConfigurable: class {
     ///   - animated: should update position and size animated
     func indicator(position: CGFloat, size: CGFloat, animated: Bool)
     
-    weak var titleViewConfigurationDelegate: TitleViewConfigurationDelegate? { get set }
+    /// Return `true` if sliding indicator should update position animated
+    /// - Parameters:
+    ///   - index: target index for sliding indicator
+    func shouldAnimateIndicatorOnSelection(index: Int) -> Bool
+    
+    var titleViewConfigurationDelegate: TitleViewConfigurationDelegate? { get set }
 }
 
 public protocol TitleViewConfigurationDelegate: class {
@@ -43,15 +48,15 @@ open class TitleScrollView<T>: UIScrollView, ViewSlidable, TitleConfigurable whe
     public typealias View = T
     public typealias TitleItem = View
     public private(set) var isLayouted = false
-    private var previousSize = CGSize.zero
-    private var previousContentSize = CGSize.zero
+    private var previousSize: CGSize = .zero
+    private var previousContentSize: CGSize = .zero
     
     public init() {
-        super.init(frame: CGRect.zero)
-        showsVerticalScrollIndicator = false
-        showsHorizontalScrollIndicator = false
+        super.init(frame: .zero)
+        self.showsVerticalScrollIndicator = false
+        self.showsHorizontalScrollIndicator = false
         if #available(iOS 11.0, *) {
-            contentInsetAdjustmentBehavior = .never
+            self.contentInsetAdjustmentBehavior = .never
         }
     }
     
@@ -61,15 +66,19 @@ open class TitleScrollView<T>: UIScrollView, ViewSlidable, TitleConfigurable whe
     
     override open func layoutSubviews() {
         super.layoutSubviews()
-        if !isLayouted {
-            isLayouted = true
-            firstLayoutAction?()
+        
+        if !self.isLayouted {
+            self.isLayouted = true
+            self.firstLayoutAction?()
         }
-        if bounds.size != previousSize || contentSize != previousContentSize {
-            previousSize = bounds.size
-            previousContentSize = contentSize
-            changeLayoutAction?()
+        
+        guard self.bounds.size != self.previousSize || self.contentSize != self.previousContentSize else {
+            return
         }
+
+        self.previousSize = self.bounds.size
+        self.previousContentSize = self.contentSize
+        self.changeLayoutAction?()
     }
     
     // MARK: - ViewSlidableImplementation
@@ -89,10 +98,10 @@ open class TitleScrollView<T>: UIScrollView, ViewSlidable, TitleConfigurable whe
     
     
     /// Alignment of title view. Supports `.top`, `.bottom`, `.left`, `.right`. The default value of `alignment` is `.top`.
-    public var alignment = TitleViewAlignment.top {
+    public var alignment: TitleViewAlignment = .top {
         didSet {
-            if alignment != oldValue {
-               titleViewConfigurationDelegate?.didChangeAlignment(alignment: alignment)
+            if self.alignment != oldValue {
+               self.titleViewConfigurationDelegate?.didChangeAlignment(alignment: self.alignment)
             }
         }
     }
@@ -100,26 +109,30 @@ open class TitleScrollView<T>: UIScrollView, ViewSlidable, TitleConfigurable whe
     /// The size of `TitleScrollView`. For `.horizontal` slide direction of `SlideController` the `titleSize` corresponds to `height`. For `.vertical` slide direction of `SlideController` the `titleSize` corresponds to `width`.  The default value of `titleSize` is `84`.
     open var titleSize: CGFloat = 84 {
         didSet {
-            if titleSize != oldValue {
-                titleViewConfigurationDelegate?.didChangeTitleSize(size: titleSize)
+            if self.titleSize != oldValue {
+                self.titleViewConfigurationDelegate?.didChangeTitleSize(size: self.titleSize)
             }
         }
     }
     
-    open var position = TitleViewPosition.beside {
+    open var position: TitleViewPosition = .beside {
         didSet {
-            if position != oldValue {
-                titleViewConfigurationDelegate?.didChangePosition(position: position)
+            if self.position != oldValue {
+                self.titleViewConfigurationDelegate?.didChangePosition(position: self.position)
             }
         }
     }
     
     open func indicator(position: CGFloat, size: CGFloat, animated: Bool) { }
     
+    open func shouldAnimateIndicatorOnSelection(index: Int) -> Bool {
+        return false
+    }
+    
     weak public var titleViewConfigurationDelegate: TitleViewConfigurationDelegate?
     
     /// Array of title items that displayed in `TitleScrollView`.
     open var items: [TitleItem] {
-        return [TitleItem]()
+        return []
     }
 }